完成算数指令

+++ ch05/src/main/java/indi/pancras/jvm/instruction/Instruction.java
@@ -11,7 +11,7 @@ public interface Instruction {
      *
      * @return 指令值
      */
-    byte getOpCode();
+    int getOpCode();
 
     /**
      * 获取指令助记符
