+++ ch05/src/main/java/indi/pancras/jvm/instruction/load/Dload1.java
@@ -8,7 +8,7 @@ import indi.pancras.jvm.rtda.Frame;
  */
 public class Dload1 extends BaseNoOperands {
     @Override
-    public byte getOpCode() {
+    public int getOpCode() {
         return 0x27;
     }
