+++ ch05/src/main/java/indi/pancras/jvm/instruction/store/Astore.java
@@ -9,7 +9,7 @@ import indi.pancras.jvm.rtda.Reference;
  */
 public class Astore extends BaseIndex8 {
     @Override
-    public byte getOpCode() {
+    public int getOpCode() {
         return 0x3a;
     }
