+++ ch05/src/main/java/indi/pancras/jvm/instruction/store/FAstore.java
@@ -8,7 +8,7 @@ import indi.pancras.jvm.rtda.Frame;
  */
 public class FAstore extends BaseIndex8 {
     @Override
-    public byte getOpCode() {
+    public int getOpCode() {
         return 0x51;
     }
